### 2024-10-09 #664 @WiIIson
onmaRg package for the Ontario Marginalization Index
Submitting Author Name: D. William Conley
Submitting Author Github Handle: @Willson\
Repository:  https://github.com/WiIIson/onmaRg
Submission type: Pre-submission

---

```
Package: onmaRg
Title: Import Public Health Ontario's Ontario Marginalization Index
Version: 1.0.3
    person("William", "Conley", , "william@cconley.ca", role = c("aut", "cre"))
Description: The Ontario Marginalization Index is a socioeconomic model that is built on Statistics Canada census data.
    The model consists of four dimensions: In 2021, these dimensions were updated to "Material Resources" (previously called "Material Deprivation"), "Households and Dwellings" (previously called "Residential Instability"), "Age and Labour Force" (previously called "Dependency"), and "Racialized and Newcomer Populations" (previously called "Ethnic Concentration").
    This update reflects a movement away from deficit-based language. 2021 data will load with these new dimension names, wheras 2011 and 2016 data will load with the historical dimension names.
    Each of these dimensions are imported for a variety of geographic levels (DA, CD, etc.) for the 2021, 2011 and 2016 administrations of the census.
    These data sets contribute to community analysis of equity with respect to Ontario's Anti-Racism Act.
    The Ontario Marginalization Index data is retrieved from the Public Health Ontario website: <https://www.publichealthontario.ca/en/data-and-analysis/health-equity/ontario-marginalization-index>.
    The shapefile data is retrieved from the Statistics Canada website: <https://www12.statcan.gc.ca/census-recensement/2011/geo/bound-limit/bound-limit-eng.cfm>.
Depends: dplyr, httr, readxl, sf, stringr, utils
```

The onmaRg package retrieves data from Public Health Ontario's Ontario Marginalization Index, and joins it with Stats Canada's geospatial data to create a geographic map of OnMarg data.

N/A.

The target audience of this package is social scientists and epidemiologists who are engaged in understanding community socioeconomic dynamics. This makes disparate datasets more accessible. There is currently a user group within education research that has been using this package.

No.

OnMarg is a data model developed by Public Health Ontario, using data from Statistics Canada.

No.
