### 2024-10-09 #664 @adamhsparks
@WiIIson, I don't understand how the package works if you can't fetch the data due to timeouts?

But if you're experiencing issues with this, I'd set the `TIMEOUT` to be longer, *e.g.*, https://github.com/ropensci/weatherOz/blob/0c544acd4651fa16a3a3dc2e3e24addf5d41187d/R/get_radar_imagery.R#L49.

I see from the bot's checks that you have a few other issues that should be addressed as well, such as the URLs in the DESCRIPTION a COC and codemeta.json file.
