### 2024-10-10 #664 @WiIIson
I'd be up for getting the testing ready, I'm just wondering if you think this package would be a good fit for OpenSci.

The tests timing out is related to CRAN accepting the package, I believe they have a fixed amount of time the tests can run for even though the functions all work properly. If the tests are running locally then it won't be a problem.
