### 2024-10-10 #664 @adamhsparks
Ah. Ok!

Yes, this would fit with data retrieval.

And have we got a book for you! https://books.ropensci.org/http-testing/

To be brief, there are packages that allow you to record the API’s response, _e.g.,_ {httptest2} or {vcr}, which make the tests much faster and you can also just use `test that::skip_if_ofline()` and the tests are skipped on CRAN. See, https://github.com/ropensci/weatherOz/blob/0c544acd4651fa16a3a3dc2e3e24addf5d41187d/tests/testthat/test-get_dpird_summaries.R#L104 for using a {vcr} cassette and skipping offline.

We would suggest updating to {httr2} as well. You’ll have a more recent and better supported package to work with.
